// src/types/DataBase/Learn.ts
interface learnDataBase {
    email: string;
    grade: string;
    volume: string;
    chineseTextbooks: string;
    mathematicsTextbooks: string;
    englishTextbooks: string;
    physicsTextbooks: string;
    chemistryTextbooks: string;
    historyTextbooks: string;
    biologyTextbooks: string;
    geographyTextbooks: string;
    morality_and_rule_of_lawTextbooks: string;
}

export {learnDataBase}